Allow versioned snapshot builds to take place
All assumptions that the latest tag is the current version of the
package have been stripped out.  The version can now be either
calculated using `au-rebuild-changelog --version`, or parsed from the
resulting changelog.

I also removed `au-mk-activate` and `au-checkout-latest-tag` as the were
only called from inside the generated Makefile.  At the same time I
tidied up the variables in the Makefile to be more obvious (to me).